Move code in saito_controller into main.rs #36
Moved code from the saito_controller.rs file into the main.rs and removed the saito_controller.rs file. Thread creation broken into separate methods.
