+++ saito-rust/src/saito/mod.rs
@@ -5,5 +5,4 @@ pub mod io_event;
 mod io_future;
 pub mod rust_io_handler;
 mod rust_task_runner;
-pub mod saito_controller;
-mod time_keeper;
+pub mod time_keeper;
